Fixes #24 Errant Dialog display on Safari
This commit attempts to mitigate the display issues
associated with using an old version of Safari browser.
Old versions of Safari do not support the <dialog> element.
This results in the dialog contents accidentally displaying on
initial load.  This fix mitgates (but does not fully solve) the
problem.

If the browser does not support the <dialog> element, the
site will fall back to "Read Only" mode, only displaying
information, not presenting UI to update information.
This is because write operations almost all utilize dialog
interactions.

Since the current version of the Safari browser >= 15.4
supports the <dialog> element, this choice was taken instead
of polyfilling the dialog functionality with an additional
javascript framework/dependency.  Chrome based browsers
fully support <dialog> at this time.