Fixes #19 Improved Error Handling
Added Error Feedback to the create
proposal page.  The Title and Voting
period are required, if the user attempts
to publish a proposal without having
entered values for these fields, the
UI will outline the fields in red and
display error text.

Additionally, if user has not yet
connected a wallet, the user will be
also prompted to connect a wallet if
they invoke the publish button.

Apologies for other small tweaks to
code, that was done by the linter.