Incorporate Voting Threshold and Ineligible List
Incorporated server support for minimum quorum
threshold setting (as a fraction of the eligible voting
balance) and list of ineligible accounts (accounts that
may not participate in a ballot vote, such as treasuries).

The server respects the threshold and ineligible list as
defined in the proposal definition, it provides global
settings to client software as an advisory for how new
proposals should be created.  However the server
software only reads the stream and does not
participate in creation of proposals therefore cannot
control the threshold and lists established when a
proposal is created.  In the future, the server code
could change to invalidate proposals that do not meet
certain criteria, but those features are not included in
this revision.