Update Website UI
Updated the website UI subproject to
support the new aditional properties
of global configuration.  Ballots created
by this user interface will conform to the
constraints defined in the 'rules-definition'
message which at message 1 in the HCS
voting stream.

If connected via hash-connect, the site will
also check the payer ID against the allowed
ballot creator's list and produce an error if
the connected payer is not included on the
ballot creators' list (if it exists in the
configuration)