Fixed Validation CLI defect
Fixed a defect in the CLI validation algorithms
that failed to account for accounts holding a
large variety of different tokens.  Adjusted the
algorithm to be more specific when querying
token balances to eliminate the possiblity of
the mirror node paging results, which was the
underlying cause of the issue.